requirement.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
ccxt==1.73.17
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
frozenlist==1.3.0
haversine==2.5.1
idna==3.3
multidict==6.0.2
numpy==1.22.2
pandas==1.4.1
pycares==4.1.2
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
pyupbit==0.2.23
requests==2.27.1
schedule==1.1.0
six==1.16.0
telegram==0.0.1
urllib3==1.26.8
websockets==9.1
xmltodict==0.13.0
yarl==1.7.2
fredapi==0.5.0